feat(Accordion): make clickable in PM, cleaned up code
- Removed circular dependency
- useCallback around accordion setting functions
- "toggle all" can be clicked in PM
