+++ components/ui/Table/AccordionController.js
@@ -1,33 +1,9 @@
-import React, { useState, useContext, useMemo } from 'react';
+import React, { useState, useMemo, useCallback } from 'react';
 
 import { useModes } from '@livepreso/content-react';
+import { AccordionControlContext } from './use-accordion-controls';
 
 import { AccordionToggleAll } from './AccordionToggleAll';
-
-export const AccordionControlContext = React.createContext({
-  registerRow: () => {},
-  unregisterRow: () => {},
-  isOpen: () => false,
-  toggleRow: () => {},
-  hasRows: false,
-  expandAll: () => {},
-  collapseAll: () => {},
-  allRowsOpen: false,
-  hasController: false,
-});
-
-export const useAccordionControls = () => {
-  const context = useContext(AccordionControlContext);
-
-  if (!context.hasController) {
-    throw new Error(`
-      useAccordionControls must be used within an AccordionController.
-      Please wrap your table with AccordionController to use this hook.
-    `);
-  }
-
-  return context;
-};
 
 export function AccordionController({ children }) {
   const { isPdfScreenshot } = useModes();
@@ -35,50 +11,62 @@ export function AccordionController({ children }) {
   const [exceptions, setExceptions] = useState(new Set());
   const [isAllOpen, setIsAllOpen] = useState(isPdfScreenshot);
 
-  const isRowExpanded = (uid) => {
-    // If Global is OPEN, an exception means CLOSED.
-    // If Global is CLOSED, an exception means OPEN.
-    return isAllOpen ? !exceptions.has(uid) : exceptions.has(uid);
-  };
+  const isRowExpanded = useCallback(
+    (uid) => {
+      // If Global is OPEN, an exception means CLOSED.
+      // If Global is CLOSED, an exception means OPEN.
+      return isAllOpen ? !exceptions.has(uid) : exceptions.has(uid);
+    },
+    [isAllOpen, exceptions],
+  );
 
   const hasRows = registry.size > 0;
   const allRowsOpen =
     (isAllOpen && exceptions.size === 0) ||
     (!isAllOpen && exceptions.size === registry.size);
 
-  const registerRow = (uid) => {
-    setRegistry((prev) => new Set(prev).add(uid));
-  };
+  const registerRow = useCallback(
+    (uid) => {
+      setRegistry((prev) => new Set(prev).add(uid));
+    },
+    [setRegistry],
+  );
 
-  const unregisterRow = (uid) => {
-    setRegistry((prev) => {
-      const newSet = new Set(prev);
-      newSet.delete(uid);
-      return newSet;
-    });
-  };
-
-  const toggleRow = (uid) => {
-    setExceptions((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(uid)) {
+  const unregisterRow = useCallback(
+    (uid) => {
+      setRegistry((prev) => {
+        const newSet = new Set(prev);
         newSet.delete(uid);
-      } else {
-        newSet.add(uid);
-      }
-      return newSet;
-    });
-  };
+        return newSet;
+      });
+    },
+    [setRegistry],
+  );
 
-  const expandAll = () => {
+  const toggleRow = useCallback(
+    (uid) => {
+      setExceptions((prev) => {
+        const newSet = new Set(prev);
+        if (newSet.has(uid)) {
+          newSet.delete(uid);
+        } else {
+          newSet.add(uid);
+        }
+        return newSet;
+      });
+    },
+    [setExceptions],
+  );
+
+  const expandAll = useCallback(() => {
     setIsAllOpen(true);
     setExceptions(new Set());
-  };
+  }, [setIsAllOpen, setExceptions]);
 
-  const collapseAll = () => {
+  const collapseAll = useCallback(() => {
     setIsAllOpen(false);
     setExceptions(new Set());
-  };
+  }, [setIsAllOpen, setExceptions]);
 
   const value = useMemo(
     () => ({
